% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_CoordFix.R
\name{launch_app}
\alias{launch_app}
\title{Launch Shiny App to convert Longitude and Latitude degrees to decimal degrees}
\usage{
launch_app()
}
\description{
Launch Shiny App to convert Longitude and Latitude degrees to decimal degrees
}
